rename: SubCategoryTab -> TableOfContents로 이름 변경
